// frontend/quickeats-web/src/app/core/models/checkout.model.ts
export interface CheckoutModel {

    
    deliveryAddress: string;

    
    landmark: string;

    phoneNumber: string;

    
    deliveryInstruction: string;

   
    paymentMethod: string;

    couponCode: string;

    subTotal: number;

    gstAmount: number;

    deliveryCharge: number;

    platformFee: number;

    discountAmount: number;

    grandTotal: number;

    rewardPoints: number;

}